+++ modules/aws/alerting/main.tf
@@ -32,7 +32,7 @@ locals {
 
 module "apres_names" {
   #checkov:skip=CKV_TF_1:False positive, we are not using a hash because we use the tagged version.
-  source      = "git@github.com:apresdev/apres-terraform.git//modules/aws/apres_names?ref=rel/apres_names/1.0.0"
+  source      = "https://github.com/apresdev/apres-terraform.git//modules/aws/apres_names?ref=rel/apres_names/2.0.1"
   name        = var.name
   environment = var.environment
 }