+++ modules/aws/ecs/tests/fixtures/main.tf
@@ -35,7 +35,7 @@ module "ecs" {
   environment = var.environment
   # This container is built by apresdev/terraform-test-support and is a simple
   # web server with a static page.
-  container_image_uri           = "767397774077.dkr.ecr.us-east-2.amazonaws.com/apres-hello-world:6333ff88fed52c2b7c531f35a56af28e21ec4ec5"
+  container_image_uri           = "123456789012.dkr.ecr.us-east-2.amazonaws.com/apres-hello-world:6333ff88fed52c2b7c531f35a56af28e21ec4ec5"
   deployment_target             = var.target
   vpc_environment_tag           = var.vpc_environment_tag
   ec2_use_instance_nvme_storage = var.ec2_use_instance_nvme_storage