+++ modules/aws/acm_public_cert/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This change log is automatically generated.
 
+## 2.0.1 - 2026-05-22
+
+Switched source from SSH to HTTPS
+
 ## 2.0.0 - 2024-07-25
 
 Change the build process to use Makefiles.