+++ modules/aws/lambda_scheduler/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This change log is automatically generated.
 
+## 0.1.1 - 2026-05-22
+
+Switched source from SSH to HTTPS
+
 ## 0.1.0 - 2024-11-22
 
 Add a Lambda scheduler module.